Update main.f90
Fixing problem with overflowing. Using intrinsic floor. I don't if intrinsics are permited in this project because there's no guidance for contributors.

Also there is `index` intrinsic (from F77) to find first occurrence of a character in a string it could be changed but `y`?

Also corrected the length of the alphabet